Phlex form helpers honor `prefs: true` (issue #4258 item 8)
Adds the Phlex-side equivalent of ERB
`auto_label_if_form_is_account_prefs`: when callers pass `prefs: true`
to a form helper, auto-resolve the label from the
`prefs_<field>` i18n key and strip the `:prefs` key before it flows
into wrapper options. Applies to the same six helpers ERB does it
for: `text_field`, `textarea_field`, `select_field`,
`checkbox_field`, `radio_field`, `number_field`.

Implementation: small private helper `auto_label_for_prefs(field_name,
options)` on `ApplicationForm`, called as the first line of each of
the 6 public helpers. Direct mirror of the ERB shape, so the i18n
keys MO already maintains (`config/locales/en.txt`:
`prefs_login`, `prefs_no_emails`, etc.) work identically across
both rendering paths.

## Why two extracted modules in this PR

Adding the prefs helper pushed the `application_form.rb` class over
the rubocop `Metrics/ClassLength` limit (250 lines). Rather than
disable the cop, extract the same chunks PR #4250 extracts:

- `application_form/field_proxy.rb`: hosts `FieldProxy` + inner
  `DOMProxy` (self-contained, used by FieldProxy-backed field
  rendering outside form context).
- `application_form/upload_helpers.rb`: hosts `upload_fields`,
  `image_namespace`, and the private `render_upload_*` renderers
  (only relevant to forms that handle image uploads).

Both extracts match the file structure PR #4250 introduces, so when
both eventually merge they'll align cleanly. Zeitwerk autoloads
both.

## Tests

- 57 tests / 302 assertions pass in `application_form_test.rb`,
  including new regressions:
  - `test_text_field_prefs_auto_resolves_label_from_i18n` —
    `prefs_login` → "Login" appears in the rendered label.
  - `test_checkbox_field_prefs_auto_resolves_label_from_i18n` —
    `prefs_no_emails` → "Opt out of …" appears in the rendered label.
  - `test_auto_label_for_prefs_returns_options_unchanged_when_no_prefs`
    — passthrough when `:prefs` is absent.
  - `test_auto_label_for_prefs_drops_prefs_key_when_resolving` —
    confirms `:prefs` is removed after resolution so it doesn't leak
    into wrapper_options downstream.
- Upload-helper-using form tests (glossary_term_form, account_profile_form,
  project_form, plus observation/herbarium/name/sequence_form) all
  pass — confirms the upload-helpers module extraction is behaviorally
  invisible.
- `bundle exec rubocop` clean.
